requirements.test.txt
colorlog>=6.10.1
homeassistant>=2025.10.3
pip>=25.2
ruff>=0.14.1
pytest-homeassistant-custom-component>=0.13.288